Extract createdb methods out of pgx-utils
