+++ cargo-pgx/src/command/init.rs
@@ -13,8 +13,7 @@ use crate::CommandExecute;
 use eyre::{eyre, WrapErr};
 use owo_colors::OwoColorize;
 use pgx_paths::prefix_path;
-use pgx_pg_config::{PgConfig, PgConfigSelector, Pgx};
-use pgx_utils::SUPPORTED_MAJOR_VERSIONS;
+use pgx_pg_config::{PgConfig, PgConfigSelector, Pgx, SUPPORTED_MAJOR_VERSIONS};
 use rayon::prelude::*;
 
 use std::collections::HashMap;
